MULE-16174: Refactor of ProcessingStrategies
