+++ core/src/main/java/org/mule/runtime/core/internal/processor/strategy/WorkQueueStreamProcessingStrategyFactory.java
@@ -6,13 +6,25 @@
  */
 package org.mule.runtime.core.internal.processor.strategy;
 
+import static java.lang.System.currentTimeMillis;
+import static java.lang.Thread.currentThread;
+import static java.time.Duration.ofMillis;
 import static java.util.Objects.requireNonNull;
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.mule.runtime.core.api.processor.ReactiveProcessor.ProcessingType.BLOCKING;
 import static org.mule.runtime.core.api.processor.ReactiveProcessor.ProcessingType.CPU_LITE_ASYNC;
 import static org.mule.runtime.core.internal.context.thread.notification.ThreadNotificationLogger.THREAD_NOTIFICATION_LOGGER_CONTEXT_KEY;
+import static org.mule.runtime.core.internal.processor.strategy.WorkQueueStreamProcessingStrategyFactory.WaitStrategy.valueOf;
 import static reactor.core.publisher.Flux.from;
 import static reactor.core.publisher.Flux.just;
 import static reactor.core.scheduler.Schedulers.fromExecutorService;
+import static reactor.util.concurrent.WaitStrategy.blocking;
+import static reactor.util.concurrent.WaitStrategy.busySpin;
+import static reactor.util.concurrent.WaitStrategy.liteBlocking;
+import static reactor.util.concurrent.WaitStrategy.parking;
+import static reactor.util.concurrent.WaitStrategy.phasedOffLiteLock;
+import static reactor.util.concurrent.WaitStrategy.sleeping;
+import static reactor.util.concurrent.WaitStrategy.yielding;
 
 import org.mule.runtime.api.exception.MuleException;
 import org.mule.runtime.api.lifecycle.Disposable;
@@ -21,16 +33,26 @@ import org.mule.runtime.api.lifecycle.Stoppable;
 import org.mule.runtime.api.scheduler.Scheduler;
 import org.mule.runtime.api.scheduler.SchedulerService;
 import org.mule.runtime.core.api.MuleContext;
+import org.mule.runtime.core.api.construct.FlowConstruct;
+import org.mule.runtime.core.api.event.CoreEvent;
 import org.mule.runtime.core.api.processor.ReactiveProcessor;
 import org.mule.runtime.core.api.processor.Sink;
 import org.mule.runtime.core.api.processor.strategy.ProcessingStrategy;
 import org.mule.runtime.core.internal.context.thread.notification.ThreadLoggingExecutorServiceDecorator;
+import org.mule.runtime.core.privileged.event.BaseEventContext;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
 
+import reactor.core.publisher.FluxSink;
 import reactor.core.publisher.Mono;
+import reactor.core.publisher.WorkQueueProcessor;
 
 /**
  * Creates {@link WorkQueueStreamProcessingStrategy} instances that de-multiplexes incoming messages using a ring-buffer but
@@ -42,7 +64,9 @@ import reactor.core.publisher.Mono;
  *
  * @since 4.0
  */
-public class WorkQueueStreamProcessingStrategyFactory extends AbstractStreamProcessingStrategyFactory {
+public class WorkQueueStreamProcessingStrategyFactory extends AbstractStreamWorkQueueProcessingStrategyFactory {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(WorkQueueStreamProcessingStrategyFactory.class);
 
   @Override
   public ProcessingStrategy create(MuleContext muleContext, String schedulersNamePrefix) {
@@ -64,7 +88,10 @@ public class WorkQueueStreamProcessingStrategyFactory extends AbstractStreamProc
 
   static class WorkQueueStreamProcessingStrategy extends AbstractStreamProcessingStrategy implements Startable, Stoppable {
 
+    private final int bufferSize;
+    private final Supplier<Scheduler> ringBufferSchedulerSupplier;
     private final Supplier<Scheduler> blockingSchedulerSupplier;
+    private final WaitStrategy waitStrategy;
     private Scheduler blockingScheduler;
     private final List<Sink> sinkList = new ArrayList<>();
     private final boolean isThreadLoggingEnabled;
@@ -74,8 +101,11 @@ public class WorkQueueStreamProcessingStrategyFactory extends AbstractStreamProc
                                                 String waitStrategy, Supplier<Scheduler> blockingSchedulerSupplier,
                                                 int maxConcurrency, boolean maxConcurrencyEagerCheck,
                                                 boolean isThreadLoggingEnabled) {
-      super(ringBufferSchedulerSupplier, bufferSize, subscribers, waitStrategy, maxConcurrency, maxConcurrencyEagerCheck);
+      super(subscribers, maxConcurrency, maxConcurrencyEagerCheck);
+      this.bufferSize = requireNonNull(bufferSize);
+      this.ringBufferSchedulerSupplier = requireNonNull(ringBufferSchedulerSupplier);
       this.blockingSchedulerSupplier = requireNonNull(blockingSchedulerSupplier);
+      this.waitStrategy = valueOf(waitStrategy);
       this.isThreadLoggingEnabled = isThreadLoggingEnabled;
     }
 
@@ -85,6 +115,43 @@ public class WorkQueueStreamProcessingStrategyFactory extends AbstractStreamProc
                                                 int maxConcurrency, boolean maxConcurrencyEagerCheck) {
       this(ringBufferSchedulerSupplier, bufferSize, subscribers, waitStrategy, blockingSchedulerSupplier, maxConcurrency,
            maxConcurrencyEagerCheck, false);
+    }
+
+    @Override
+    public Sink createSink(FlowConstruct flowConstruct, ReactiveProcessor function) {
+      final long shutdownTimeout = flowConstruct.getMuleContext().getConfiguration().getShutdownTimeout();
+      WorkQueueProcessor<EventWrapper> processor =
+          WorkQueueProcessor.<EventWrapper>builder().executor(ringBufferSchedulerSupplier.get()).bufferSize(bufferSize)
+              .waitStrategy(waitStrategy.getReactorWaitStrategy()).build();
+      int subscriberCount = maxConcurrency < subscribers ? maxConcurrency : subscribers;
+      CountDownLatch completionLatch = new CountDownLatch(subscriberCount);
+      for (int i = 0; i < subscriberCount; i++) {
+        processor
+            .doOnSubscribe(subscription -> currentThread().setContextClassLoader(executionClassloader))
+            .map(ew -> ew.getWrappedEvent())
+            .transform(function)
+            .subscribe(null, e -> completionLatch.countDown(), completionLatch::countDown);
+      }
+      return buildSink(processor.sink(), () -> {
+        long start = currentTimeMillis();
+        if (!processor.awaitAndShutdown(ofMillis(shutdownTimeout))) {
+          LOGGER.warn("WorkQueueProcessor of ProcessingStrategy for flow '{}' not shutDown in {} ms. Forcing shutdown...",
+                      flowConstruct.getName(), shutdownTimeout);
+          processor.forceShutdown();
+        }
+        awaitSubscribersCompletion(flowConstruct, shutdownTimeout, completionLatch, start);
+      }, createOnEventConsumer(), bufferSize);
+    }
+
+    protected <E> ReactorSink<E> buildSink(FluxSink<E> fluxSink, reactor.core.Disposable disposable,
+                                           Consumer<CoreEvent> onEventConsumer, int bufferSize) {
+      return new DefaultReactorSink(fluxSink, disposable, onEventConsumer, bufferSize) {
+
+        @Override
+        public EventWrapper intoSink(CoreEvent event) {
+          return new EventWrapper(event);
+        }
+      };
     }
 
     @Override
@@ -131,6 +198,48 @@ public class WorkQueueStreamProcessingStrategyFactory extends AbstractStreamProc
       }
     }
 
+  }
+
+  private static final class EventWrapper {
+
+    CoreEvent wrappedEvent;
+
+    public EventWrapper(CoreEvent event) {
+      this.wrappedEvent = event;
+      ((BaseEventContext) (wrappedEvent.getContext())).getRootContext().onTerminated((e, t) -> {
+        wrappedEvent = null;
+      });
+    }
+
+    public CoreEvent getWrappedEvent() {
+      return wrappedEvent;
+    }
+  }
+
+  protected enum WaitStrategy {
+    BLOCKING(blocking()),
+
+    LITE_BLOCKING(liteBlocking()),
+
+    SLEEPING(sleeping()),
+
+    BUSY_SPIN(busySpin()),
+
+    YIELDING(yielding()),
+
+    PARKING(parking()),
+
+    PHASED(phasedOffLiteLock(200, 100, MILLISECONDS));
+
+    private final reactor.util.concurrent.WaitStrategy reactorWaitStrategy;
+
+    WaitStrategy(reactor.util.concurrent.WaitStrategy reactorWaitStrategy) {
+      this.reactorWaitStrategy = reactorWaitStrategy;
+    }
+
+    reactor.util.concurrent.WaitStrategy getReactorWaitStrategy() {
+      return reactorWaitStrategy;
+    }
   }
 
 }
